+++ wdt_ct/w8760_funcs.cpp
@@ -727,21 +727,22 @@ int wh_w8760_prepare_data(WDT_DEV* pdev, BOARD_INFO* p_out_board_info)
 	if (p_out_board_info->dev_type & FW_WDT8760_2_ISP)
 		return 1;
 
-	if (wh_w8760_dev_get_context(pdev, &pct_data)) {
-		/* set the default values */
-		p_out_board_info->sys_param.Phy_Frmbuf_W= pct_data.n_cs;
-		p_out_board_info->sys_param.Phy_X0= pct_data.x1;
-		p_out_board_info->sys_param.Phy_X1= pct_data.xn;
 
-		p_out_board_info->sys_param.Phy_Frmbuf_H= pct_data.n_cd;
-		p_out_board_info->sys_param.Phy_Y0= pct_data.y1;
-		p_out_board_info->sys_param.Phy_Y1= pct_data.yn;
+	p_out_board_info->sys_param.Phy_Frmbuf_W= 0;
+	p_out_board_info->sys_param.Phy_X0= 0;
+	p_out_board_info->sys_param.Phy_X1= 0;
+
+	p_out_board_info->sys_param.Phy_Frmbuf_H= 0;
+	p_out_board_info->sys_param.Phy_Y0= 0;
+	p_out_board_info->sys_param.Phy_Y1= 0;
 		
-		p_out_board_info->sys_param.xmls_id1 = 0;
-		p_out_board_info->sys_param.xmls_id2 = 0;
+	p_out_board_info->sys_param.xmls_id1 = 0;
+	p_out_board_info->sys_param.xmls_id2 = 0;
+	/*
+         * We don’t need to retrieve the checksum or device PHY data when querying the device. Removing these will help speed up the query time.
+        */
 
-		return 1;
-	}
+
 
 	return 1;
 }